move new functions into step-checker/utils
