+++ src/step-checker/__tests__/util.test.ts
@@ -1,4 +1,9 @@
-import {primeDecomp} from "../util";
+import * as Util from "../../semantic/util";
+
+import {primeDecomp, findNodeById, replaceNodeWithId} from "../util";
+
+import serializer from "../../semantic/semantic-serializer";
+expect.addSnapshotSerializer(serializer);
 
 describe("primeDecomp", () => {
     it("should return an empty array for decimal numbers", () => {
@@ -19,5 +24,67 @@ describe("primeDecomp", () => {
 
     it("36 -> [2, 2, 3, 3]", () => {
         expect(primeDecomp(36)).toEqual([2, 2, 3, 3]);
+    });
+});
+
+describe("findNodeById", () => {
+    test("finds the root node", () => {
+        const root = Util.number("1");
+
+        const node = findNodeById(root, root.id);
+
+        expect(node).toBe(root);
+    });
+
+    test("finds a node in an array", () => {
+        const one = Util.number("1");
+        const two = Util.number("2");
+        const root = Util.add([one, two]);
+
+        const node = findNodeById(root, two.id);
+
+        expect(node).toBe(two);
+    });
+
+    test("finds a node in named property", () => {
+        const x = Util.number("x");
+        const two = Util.number("2");
+        const root = Util.exp(x, two);
+
+        const node = findNodeById(root, two.id);
+
+        expect(node).toMatchInlineSnapshot;
+    });
+
+    test("doesn't find a node", () => {
+        const root = Util.number("1");
+
+        const node = findNodeById(root, -1);
+
+        expect(node).toBe(undefined);
+    });
+});
+
+describe("replaceNode", () => {
+    test("replaces a node in an array", () => {
+        const one = Util.number("1");
+        const two = Util.number("2");
+        const root = Util.add([one, two]);
+        const three = Util.number("3");
+
+        replaceNodeWithId(root, two.id, three);
+
+        expect(root).toMatchInlineSnapshot(`(add 1 3)`);
+    });
+
+    test("finds a node in named property", () => {
+        const x = Util.number("x");
+        const two = Util.number("2");
+        const root = Util.exp(x, two);
+        const three = Util.number("3");
+
+        replaceNodeWithId(root, two.id, three);
+
+        expect(root).toMatchInlineSnapshot(`(exp :base x :exp 3)`);
     });
 });
