+++ src/step-checker/step-checker.ts
@@ -1,10 +1,10 @@
 import BigNumber from "bignumber.js";
-import produce from "immer";
 
 import * as Semantic from "../semantic/semantic";
 import * as Util from "../semantic/util";
 
-import {zip} from "./util";
+import {zip, deepEquals, applySubReasons} from "./util";
+import {Result, Reason} from "./types";
 
 import FractionChecker from "./fraction-checker";
 import EquationChecker from "./equation-checker";
@@ -39,16 +39,6 @@ const parseNode = (node: Semantic.Expression): BigNumber => {
     }
 };
 
-export type Reason = {
-    message: string;
-    nodes: Semantic.Expression[];
-};
-
-export type Result = {
-    equivalent: boolean;
-    reasons: Reason[];
-};
-
 // TODO: fix flowtype/define-flow-type, HasArgs is used below
 // eslint-disable-next-line no-unused-vars
 type HasArgs =
@@ -72,123 +62,6 @@ export const hasArgs = (a: Semantic.Expression): a is HasArgs =>
     a.type === "gt" ||
     a.type === "gte" ||
     a.type === "div";
-
-const isNode = (val: any): val is Semantic.Expression => {
-    return Object.prototype.hasOwnProperty.call(val, "type");
-};
-
-const findNodeById = (
-    root: Semantic.Expression,
-    id: number,
-): Semantic.Expression | void => {
-    for (const val of Object.values(root)) {
-        if (isNode(val)) {
-            if (val.id === id) {
-                return val;
-            } else {
-                const result = findNodeById(val, id);
-                if (result) {
-                    return result;
-                }
-            }
-        } else if (Array.isArray(val)) {
-            for (const child of val) {
-                if (isNode(child)) {
-                    if (child.id === id) {
-                        return child;
-                    } else {
-                        const result = findNodeById(child, id);
-                        if (result) {
-                            return result;
-                        }
-                    }
-                }
-            }
-        }
-    }
-};
-
-const replaceNodeWithId = (
-    root: Semantic.Expression,
-    id: number,
-    replacement: Semantic.Expression,
-): Semantic.Expression | void => {
-    for (const val of Object.values(root)) {
-        if (isNode(val)) {
-            if (val.id === id) {
-                return val;
-            } else {
-                const result = replaceNodeWithId(val, id, replacement);
-                if (result) {
-                    return result;
-                }
-            }
-        } else if (Array.isArray(val)) {
-            for (const [index, child] of val.entries()) {
-                if (isNode(child)) {
-                    if (child.id === id) {
-                        val[index] = replacement;
-                        return child;
-                    } else {
-                        const result = replaceNodeWithId(
-                            child,
-                            id,
-                            replacement,
-                        );
-                        if (result) {
-                            if (result == child) {
-                                val[index] = replacement;
-                            }
-                            return result;
-                        }
-                    }
-                }
-            }
-        }
-    }
-};
-
-const applySubReasons = (
-    root: Semantic.Expression,
-    subreasons: Reason[],
-): Semantic.Expression => {
-    const nextState = produce(root, draft => {
-        for (const reason of subreasons) {
-            // Not all reaons come with nodes yet.
-            if (reason.nodes.length === 2) {
-                replaceNodeWithId(draft, reason.nodes[0].id, reason.nodes[1]);
-            }
-        }
-    });
-    return nextState;
-};
-
-const deepEquals = (a: any, b: any): boolean => {
-    if (Array.isArray(a) && Array.isArray(b)) {
-        return (
-            a.length === b.length &&
-            a.every((val, index) => deepEquals(val, b[index]))
-        );
-    } else if (
-        typeof a === "object" &&
-        a != null &&
-        typeof b === "object" &&
-        b != null
-    ) {
-        const aKeys = Object.keys(a).filter(key => key !== "id");
-        const bKeys = Object.keys(b).filter(key => key !== "id");
-        if (aKeys.length !== bKeys.length) {
-            return false;
-        }
-        return aKeys.every(
-            key =>
-                Object.prototype.hasOwnProperty.call(b, key) &&
-                deepEquals(a[key], b[key]),
-        );
-    } else {
-        return a === b;
-    }
-};
 
 // TODO: write a function to determine if an equation is true or not
 // e.g. 2 = 5 -> false, 5 = 5 -> true
@@ -451,10 +324,6 @@ class StepChecker implements IStepChecker {
                 [right, left],
             ]) {
                 if (y.type === "add" && y.args.length === addNode.args.length) {
-                    // TODO: use exactMatch instead here... or we'll have track all
-                    // of the reasons that are generated
-                    // TODO: apply the subReasons to previous node to get the node
-                    // before next.
                     const subReasons: Reason[] = [];
                     const equivalent = addNode.args.every((arg, index) => {
                         // Each term is in the correct order based on whether
